Improve handling of global and local locks to mitigate capture deadlocks
