+++ pkg/capture/capture.go
@@ -44,57 +44,6 @@ var (
 // providing the ability to override the default behavior, e.g. in mock tests
 type sourceInitFn func(*Capture) (Source, error)
 
-// Captures denotes a named set of Capture instances, wrapping a map and the
-// required synchronization of all its actions
-type captures struct {
-	Map map[string]*Capture
-	sync.RWMutex
-}
-
-// newCaptures instantiates a new, empty set of Captures
-func newCaptures() *captures {
-	return &captures{
-		Map:     make(map[string]*Capture),
-		RWMutex: sync.RWMutex{},
-	}
-}
-
-// Ifaces return the list of names of all interfaces in the set
-func (c *captures) Ifaces(ifaces ...string) []string {
-	if len(ifaces) == 0 {
-		c.RLock()
-		ifaces = make([]string, 0, len(c.Map))
-		for iface := range c.Map {
-			ifaces = append(ifaces, iface)
-		}
-		c.RUnlock()
-	}
-
-	return ifaces
-}
-
-// Get safely returns a Capture by name (and an indicator if it exists)
-func (c *captures) Get(iface string) (capture *Capture, exists bool) {
-	c.RLock()
-	capture, exists = c.Map[iface]
-	c.RUnlock()
-	return
-}
-
-// Set safely adds / overwrites a Capture by name
-func (c *captures) Set(iface string, capture *Capture) {
-	c.Lock()
-	c.Map[iface] = capture
-	c.Unlock()
-}
-
-// Delete safely removes a Capture from the set by name
-func (c *captures) Delete(iface string) {
-	c.Lock()
-	delete(c.Map, iface)
-	c.Unlock()
-}
-
 // Capture captures and logs flow data for all traffic on a
 // given network interface. For each Capture, a goroutine is
 // spawned at creation time. To avoid leaking this goroutine,
@@ -131,6 +80,11 @@ type Capture struct {
 
 	// startedAt tracks when the capture was started
 	startedAt time.Time
+
+	// Mutex to allow concurrent access to capture components
+	// This is _unrelated_ to the three-point capture lock to
+	// interrupt the capture for purposes of e.g. rotation
+	sync.Mutex
 }
 
 // newCapture creates a new Capture associated with the given iface.
@@ -177,6 +131,15 @@ func (c *Capture) run(memPool *LocalBufferPool) (err error) {
 }
 
 func (c *Capture) close() error {
+
+	// Lock the whole capture to pretect against duplicate close() calls
+	c.Lock()
+	defer c.Unlock()
+
+	// in case the captureHandle is already closed, return without action
+	if c.captureHandle == nil {
+		return nil
+	}
 	if err := c.captureHandle.Close(); err != nil {
 		return err
 	}
